Update patch set 4
Patch Set 4: Code-Review+1

I fixed a few mistakes (bad indention) and added a few smaller updates. I tried to keep it simple, but there is just a lot going on here.
* Removed an unused function.
* Fixed a wrong return type and made others more specific.
* Made use of the ?? operator to simplify some hard to read code.
* Removed pointless comments that repeat exactly what the code already says.

I tried hard to keep the diff readable. To make this not blow up I had to exclude another sniff. This is just temporary. We are (obviously) going to fix this very soon in the next patch.

Please feel free to +2 this if you are fine with these additions. Or give a +1 and I will +2 it.

Patch-set: 4
Reviewer: Gerrit User 1453 <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1